pyproject.toml
[project]
name = "pythondfr"
version = "0.1.0"
description = ""
authors = [
    {name = "DyakivNazar",email = "nazar2005.03.21@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "isort (>=7.0.0,<8.0.0)"
]
